pyproject.toml
[tool.poetry]
name = "shapy"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
kornia = "0.6.7"
open3d = "0.13.0"
fastapi = "0.95.2"
loguru = "0.7.0"
matplotlib = "3.7.1"
opencv-python ="4.6.0.66"
omegaconf = "2.3.0"
fvcore = "0.1.5.post20221221"
trimesh = "3.21.7"
pytorch-lightning="1.7.1"
pyrender = "0.1.45"
nflows = "0.14"
jpeg4py = "0.1.4"
torchvision="0.13.1"
torchgeometry="0.1.2"
smplx = "0.1.28"
imageio = "2.29.0"
scikit-image = "0.20.0"
scikit-learn = "1.2.2"
scipy = "1.8.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
